requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.2.2
uvicorn==0.30.6
